Moved \App\AddressFinder to \App\Map\Address
